# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pavisie API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/guilds/{guildId}/plugins": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{17,20}$"
            },
            "in": "path",
            "name": "guildId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/guilds/{guildId}/plugins/{pluginId}/config-schema": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{17,20}$"
            },
            "in": "path",
            "name": "guildId",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "moderation",
                "automod",
                "enforcer",
                "logging",
                "tickets",
                "roles",
                "engagement",
                "community",
                "gamestats",
                "economy",
                "utility",
                "media",
                "integrations",
                "ai"
              ]
            },
            "in": "path",
            "name": "pluginId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/verify/{token}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "response": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "response"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}